iniciativas-backup/app/requirements.txt
boto3>=1.33
botocore>=1.33
PyYAML>=6.0
typer[all]>=0.9
